Update README.md
In this commit, I have added the Vrinda Store Data Analysis project to the repository. The project conducts a detailed analysis of sales and inventory data for Vrinda Store using Excel. The README.md file provides an overview of the project, including its objectives, key steps, and instructions on how to explore the analysis. I have also included the Excel file containing the raw data, along with any other relevant files. Additionally, I updated the "Connect with Me" section in the README.md file with my LinkedIn profile URL for easy communication. This commit serves as the initial setup for the project on GitHub.